services/ocr-worker-py/requirements.txt
# PDF handling — in-memory rendering, no temp files
PyMuPDF==1.24.5

# OCR
pytesseract==0.3.13
Pillow==10.4.0

# Database
asyncpg==0.29.0

# Object storage
minio==7.2.9

# HTTP (webhook)
httpx==0.27.2

# File watching
watchdog==4.0.2

# Environment
python-dotenv==1.0.1
